# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Notes App API
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: number;
  title: string;
  content: string;
  createdAt: Date;
  updatedAt: Date;
}